Improve time complexity from O(n) to O(log(n))
The new algorithm searches recursively through left or right sub-lists
depending on some conditions. If the mid-element is a peak it is
returned immediately. Otherwise, if it's value is less than that of the
element to its right then that means that there must be a peak towards
the right of the mid-element (between the mid-element's index and the
index of the last element of the list). Similarly if it is less than
the element to its left then it means that there must be a peak
towards the left of the mid-element (between index 0 and the index of
the mid element).